requirements.txt
aioredis==1.2.0
amqp==2.5.0
asgiref==3.2.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
backcall==0.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
decorator==4.4.0
Django==2.2.3
django-cors-headers==3.0.2
django-filter==2.2.0
django-redis==4.10.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
dwebsocket==0.5.10
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
kombu==4.6.3
msgpack==0.6.1
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==2.0.9
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.1
redis==3.3.5
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
zope.interface==4.6.0
